Time: 353 ms (36.12%), Space: 10.9 MB (98.69%) - LeetHub

+++ 2218-maximum-value-of-k-coins-from-piles/2218-maximum-value-of-k-coins-from-piles.cpp
@@ -2,12 +2,13 @@ class Solution {
 public:
     int maxValueOfCoins(vector<vector<int>>& piles, int K) {
         vector<int> dp(K + 1, 0);
-        for (auto& pile : piles) {
-            for (int k = K; k >= 1; k--) {
+        for (auto& pile: piles) {
+            for(size_t k = K; k >= 1; k--) {
                 int pile_sum = 0;
-                for(int j = 0; j < min(k, (int)pile.size()); j++) {
+                for(size_t j = 0; j < min(k, pile.size()); j++) {
                     pile_sum += pile[j];
-                    dp[k] = max(dp[k], dp[k - j - 1] + pile_sum);
+                    dp[k] = max(dp[k], dp[ k - j - 1] + pile_sum);
+
                 }
             }
         }
